+++ CHANGELOG.md
@@ -1,5 +1,12 @@
 # Changelog
 
+## 0.0.2 (2025-11-17)
+
+### Enhancements
+
+- Add `path` option to be able to give an absolute or relative path to the web app
+- Add `openApiEndpoint` option to be able to set a custom endpoint
+
 ## 0.0.1 (2025-11-15)
 
 - Initial release